% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SIAClassifierFun.R
\name{naImpute}
\alias{naImpute}
\title{Title Gene expression data imputation}
\usage{
naImpute(Exp, G = NULL)
}
\arguments{
\item{Exp}{a dataframe with Gene Expression Profiles data values, samples in columns, genes in rows, rownames corresponding to gene symbols}

\item{G}{a character containing gene symbol(s)}
}
\value{
a Gene Expression Profiles after imputing NA values
}
\description{
Title Gene expression data imputation
}
