% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SIAClassifierFun.R
\name{SIAClassifier}
\alias{SIAClassifier}
\title{Title small intestinal adenocarcinoma molecular subtypes prediction}
\usage{
SIAClassifier(
  Expr = NULL,
  minPosterior = 0.5,
  scale = TRUE,
  log2transfrom = FALSE
)
}
\arguments{
\item{Expr}{a dataframe with Gene Expression Profiles data values, samples in columns, genes in rows, rownames corresponding to gene symbols}

\item{minPosterior}{minimal posterior probability to classify a sample. By default is 0.5.}

\item{scale}{scaling would be performed (when scale=TRUE)}

\item{log2transfrom}{log2transfrom would be performed (when log2transfrom=TRUE)}
}
\value{
a data frame with SIA subtype label and probability for each sample.
}
\description{
Title small intestinal adenocarcinoma molecular subtypes prediction
}
